Updating To JRelay2.0, New files under app/com/jr2

+++ JRelay/app/com/packets/server/AOEPacket.java
@@ -36,7 +36,7 @@ public class AOEPacket extends Packet {
 		this.pos.writeToOutput(out);
 		out.writeFloat(this.radius);
 		out.writeShort(this.damage);
-		out.writeByte(this.effect);
+		out.writeByte((byte)this.effect);
 		out.writeFloat(this.duration);
 		out.writeShort(this.origType);
 		out.writeInt(this.color);
